task176: map improvements
