+++ src/components/location-switcher/index.js
@@ -1,6 +1,5 @@
 import AppErrors from "@/js/appErrors.js";
-import { LatLng } from "leaflet";
-import { bboxes, decode_bbox } from "ngeohash";
+import { GeoHashDivider } from "@/js/geohashUtils.js";
 
 export default {
 	name: "Location",
@@ -194,8 +193,12 @@ export default {
 				const ids = this.sdk.requestServiceData.ids;
 				ids.getBrtOffersFeed += 1;
 
+				const geohashDivider = new GeoHashDivider(
+					this.map.mapObject.getBounds()
+				);
+
 				const request = {
-					location: this.getVisibleBoundsGeohashes(),
+					location: geohashDivider.getVisibleBoundsGeohashes(),
 					pageSize: this.offersRequestData.pageSize,
 					pageStart,
 					topHeight, 
@@ -236,110 +239,6 @@ export default {
 				this.setMapActionData();
 			}
 
-		},
-
-		getVisibleBoundsGeohashes() {
-			const mainBounds = this.map.mapObject.getBounds();
-			let precision = 1;
-			let result = this.getGeohashesForBounds(mainBounds, precision);
-			let temp = [];
-			const params = {
-				maxCount: 300,
-				maxPrecision: 7
-			};
-
-			while (true) {
-				result.forEach(item => {
-
-					const needDivide =
-						this.geohashCoversBounds(item, mainBounds) 
-						&& !(this.geohashIsEqualToBounds(item, mainBounds))
-						|| this.geohashIntersectsBounds(item, mainBounds);
-					
-					if (needDivide) {
-						const bounds = this.getBoundsOfGeohash(item);
-						
-						let newItems = this.getGeohashesForBounds(
-							bounds, 
-							precision + 1
-						).filter(f => !(temp.includes(f) || this.geohashOutOfBounds(f, mainBounds)));
-
-						temp = temp.concat(newItems);
-
-					} else if (this.geohashInsideBounds(item, mainBounds)) {
-						temp.push(item);
-					}
-				})
-
-				if (temp.length <= params.maxCount && precision < params.maxPrecision) {
-					result = temp;
-					temp = [];
-					precision++;
-				} else {
-					break;
-				}
-			}
-
-			return result;
-		},
-
-		getBoundsOfGeohash(geohash) {
-			const [minLat, minLon, maxLat, maxLon] = decode_bbox(geohash);
-			return { 
-				_northEast: { lat: maxLat, lng: maxLon }, 
-				_southWest: { lat: minLat, lng: minLon } 
-			}
-		},
-
-		getGeohashesForBounds(bounds, precision) {
-			return bboxes(
-				bounds._southWest.lat, 
-				bounds._southWest.lng, 
-				bounds._northEast.lat, 
-				bounds._northEast.lng, 
-				precision
-			);
-		},
-
-		geohashCoversBounds(geohash, bounds) {
-			const geohashBounds = this.getBoundsOfGeohash(geohash);
-			return geohashBounds._northEast.lat >= bounds._northEast.lat
-				&& geohashBounds._northEast.lng >= bounds._northEast.lng
-				&& geohashBounds._southWest.lat <= bounds._southWest.lat
-				&& geohashBounds._southWest.lng <= bounds._southWest.lng
-		},
-
-		geohashIsEqualToBounds(geohash, bounds) {
-			const geohashBounds = this.getBoundsOfGeohash(geohash);
-			return geohashBounds._northEast.lat === bounds._northEast.lat
-				&& geohashBounds._northEast.lng === bounds._northEast.lng
-				&& geohashBounds._southWest.lat === bounds._southWest.lat
-				&& geohashBounds._southWest.lng === bounds._southWest.lng
-		},
-
-		geohashOutOfBounds(geohash, bounds) {
-			const geohashBounds = this.getBoundsOfGeohash(geohash);
-			return geohashBounds._southWest.lat >= bounds._northEast.lat
-				|| geohashBounds._southWest.lng >= bounds._northEast.lng
-				|| geohashBounds._northEast.lat <= bounds._southWest.lat
-				|| geohashBounds._northEast.lng <= bounds._southWest.lng;
-
-		},
-
-		geohashInsideBounds(geohash, bounds) {
-			const geohashBounds = this.getBoundsOfGeohash(geohash);
-			return geohashBounds._northEast.lat <= bounds._northEast.lat
-				&& geohashBounds._northEast.lng <= bounds._northEast.lng
-				&& geohashBounds._southWest.lat >= bounds._southWest.lat
-				&& geohashBounds._southWest.lng >= bounds._southWest.lng;
-		},
-
-		geohashIntersectsBounds(geohash, bounds) {
-			return !(
-				this.geohashCoversBounds(geohash, bounds) 
-				|| this.geohashOutOfBounds(geohash, bounds)
-				|| this.geohashInsideBounds(geohash, bounds)
-			);
 		},
 
 		setMapActionData(offers, error) {
